requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.1
RealtimeSTT>=0.3.100
elevenlabs>=1.9.0
PySide6>=6.7.0
opencv-python>=4.10.0.84
Pillow>=10.4.0
mss>=9.0.1
websockets>=12.0
numpy>=2.1.1
pyaudio>=0.2.14
webrtcvad>=2.0.10
